Restructure re_arrow_combinators
Groups the transform implementations by the type of operations:

- `cast`: Transforms that cast arrays to different types.
- `map`: Transforms that apply operations to elements within arrays.
- `reshape.rs`: Transforms that extract and reshape arrays.
- `semantic.rs`: Semantic array transforms for concrete applications.
