+++ crates/utils/re_arrow_combinators/src/semantic.rs
@@ -1,0 +1,64 @@
+//! Semantic array transforms for concrete applications.
+
+use std::marker::PhantomData;
+use std::sync::Arc;
+
+use arrow::array::{Array as _, GenericBinaryArray, GenericListArray, OffsetSizeTrait};
+use arrow::datatypes::{DataType, Field};
+
+use crate::{Error, Transform};
+
+/// Converts binary arrays to list arrays where each binary element becomes a list of `u8`.
+///
+/// The underlying bytes buffer is reused, making this transformation almost zero-copy.
+#[derive(Clone, Debug, Default)]
+pub struct BinaryToListUInt8<O1: OffsetSizeTrait, O2: OffsetSizeTrait = O1> {
+    _from_offset: PhantomData<O1>,
+    _to_offset: PhantomData<O2>,
+}
+
+impl<O1: OffsetSizeTrait, O2: OffsetSizeTrait> BinaryToListUInt8<O1, O2> {
+    /// Create a new transformation to convert a binary array to a list array of `u8` arrays.
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
+
+impl<O1: OffsetSizeTrait, O2: OffsetSizeTrait> Transform for BinaryToListUInt8<O1, O2> {
+    type Source = GenericBinaryArray<O1>;
+    type Target = GenericListArray<O2>;
+
+    fn transform(&self, source: &GenericBinaryArray<O1>) -> Result<Self::Target, Error> {
+        use arrow::array::UInt8Array;
+        use arrow::buffer::ScalarBuffer;
+
+        let scalar_buffer: ScalarBuffer<u8> = ScalarBuffer::from(source.values().clone());
+        let uint8_array = UInt8Array::new(scalar_buffer, None);
+
+        // Convert from O1 to O2. Most offset buffers will be small in real-world
+        // examples, so we're fine copying them.
+        //
+        // This could be true zero copy if Rust had specialization.
+        // More info: https://std-dev-guide.rust-lang.org/policy/specialization.html
+        let old_offsets = source.offsets().iter();
+        let new_offsets: Result<Vec<O2>, Error> = old_offsets
+            .map(|&offset| {
+                let offset_usize = offset.as_usize();
+                O2::from_usize(offset_usize).ok_or_else(|| Error::OffsetOverflow {
+                    actual: offset_usize,
+                    expected_type: std::any::type_name::<O2>(),
+                })
+            })
+            .collect();
+        let offsets = arrow::buffer::OffsetBuffer::new(new_offsets?.into());
+
+        let list = Self::Target::new(
+            Arc::new(Field::new_list_field(DataType::UInt8, false)),
+            offsets,
+            Arc::new(uint8_array),
+            source.nulls().cloned(),
+        );
+
+        Ok(list)
+    }
+}
