+++ crates/utils/re_arrow_combinators/src/transform.rs
@@ -1,0 +1,54 @@
+use arrow::array::Array;
+
+use crate::Error;
+
+/// A transformation that converts one Arrow array type to another.
+///
+/// Transformations are read-only operations that may fail (e.g., missing field, type mismatch).
+/// They can be composed using the `then` method to create complex transformation pipelines.
+pub trait Transform {
+    /// The source array type.
+    type Source: Array;
+
+    /// The target array type.
+    type Target: Array;
+
+    /// Apply the transformation to the source array.
+    fn transform(&self, source: &Self::Source) -> Result<Self::Target, Error>;
+
+    /// Chain this transformation with another transformation.
+    fn then<T2>(self, next: T2) -> Compose<Self, T2>
+    where
+        Self: Sized,
+        T2: Transform<Source = Self::Target>,
+    {
+        Compose {
+            first: self,
+            second: next,
+        }
+    }
+}
+
+/// Composes two transformations into a single transformation.
+///
+/// This is the result of calling `.then()` on a transformation.
+#[derive(Clone)]
+pub struct Compose<T1, T2> {
+    first: T1,
+    second: T2,
+}
+
+impl<T1, T2, M> Transform for Compose<T1, T2>
+where
+    T1: Transform<Target = M>,
+    T2: Transform<Source = M>,
+    M: Array,
+{
+    type Source = T1::Source;
+    type Target = T2::Target;
+
+    fn transform(&self, source: &Self::Source) -> Result<Self::Target, Error> {
+        let mid = self.first.transform(source)?;
+        self.second.transform(&mid)
+    }
+}
