+++ crates/utils/re_arrow_combinators/tests/transform.rs
@@ -2,7 +2,13 @@
 
 use std::sync::Arc;
 
-use re_arrow_combinators::*;
+use re_arrow_combinators::{
+    Error, Transform as _,
+    cast::PrimitiveCast,
+    map::{MapFixedSizeList, MapList, MapPrimitive, ReplaceNull},
+    reshape::{Flatten, GetField, StructToFixedList},
+    semantic::BinaryToListUInt8,
+};
 
 use arrow::{
     array::{
